requirements.txt
fastapi
uvicorn
python-dotenv
telethon
qrcode
python-multipart
pillow
google-genai
